requirements.txt
absl-py==2.2.2
alabaster==0.7.16
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==25.3.0
babel==2.17.0
boto3==1.38.8
botocore==1.38.8
build==1.2.2.post1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.2
coverage==5.3.1
cryptography==44.0.3
cssselect2==0.8.0
cycler==0.12.1
data-gradients==0.3.2
defusedxml==0.7.1
Deprecated==1.2.18
docutils==0.17.1
einops==0.3.2
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
future==1.0.0
grpcio==1.71.0
html5lib==1.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imagededup==0.3.2
imageio==2.37.0
imagesize==1.4.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
json-tricks==3.16.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.23.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.0
onnxsim==0.4.36
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
oscrypto==1.3.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pip-tools==7.4.1
platformdirs==4.3.7
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pycocotools==2.0.8
pycparser==2.22
pyDeprecate==0.3.2
Pygments==2.19.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pyparsing==3.2.3
pypdf==5.4.0
pyproject_hooks==1.2.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
qrcode==8.2
qudida==0.0.4
RapidFuzz==3.13.0
referencing==0.36.2
reportlab==3.6.13
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
s3transfer==0.12.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
snowballstemmer==2.2.0
Sphinx==4.0.3
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stringcase==1.2.0
super-gradients==3.7.1
svglib==1.5.1
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==1.1.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tinycss2==1.4.0
tomli==2.2.1
torch==2.7.0
torchmetrics==0.8.0
torchvision==0.22.0
tqdm==4.67.1
treelib==1.6.1
triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.127
ultralytics-thop==2.0.14
uritools==5.0.0
urllib3==2.4.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
xhtml2pdf==0.2.11
